Add new color name list popover
This fancy little component allows the user to select a naming list for
their color. This is not only very fun to play around with, but also
very useful for specific tasks. The design is optimized. It is 1:17 AM
and I can't write good commit messages anymore. I apologize to whoevery
is reading this. Please stop and have some fun with the application.
Have you tried pronouncing some of these French colors?